requirements.txt
openai==1.3.0
httpx==0.24.1
tkinter
requests==2.32.3
jsonlines==3.1.0
python-dotenv==1.0.0
